requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asn1crypto==1.5.1
asttokens==2.1.0
async-timeout==4.0.3
attrs==22.1.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.11.1
bitarray==2.8.2
bleach==5.0.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.7
coincurve==18.0.0
colorama==0.4.6
comm==0.2.0
cryptography==41.0.4
cytoolz==0.12.2
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
Django==4.1.3
entrypoints==0.4
eth-abi==4.2.1
eth-account==0.10.0
eth-hash==0.5.2
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.5.2
eth-utils==2.3.1
exceptiongroup==1.1.3
executing==1.2.0
fastapi==0.104.1
fastjsonschema==2.16.2
filelock==3.12.0
frozenlist==1.4.0
functions==0.7.0
greenlet==3.0.1
h11==0.14.0
hexbytes==0.3.1
idna==3.4
iniconfig==2.0.0
ipykernel==6.20.2
ipython==8.17.2
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.3.1
json5==0.9.10
jsonschema==4.17.0
jupyter-server==1.24.0
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyterlab-pygments==0.2.2
jupyterlab_server==2.16.2
lru-dict==1.3.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mistune==2.0.4
multidict==6.0.4
nbclassic==0.4.8
nbclient==0.9.0
nbconvert==7.2.3
nbformat==5.7.0
nest-asyncio==1.5.6
notebook_shim==0.2.2
packaging==21.3
pandocfilters==1.5.0
parsimonious==0.9.0
parso==0.8.3
pickleshare==0.7.5
pip-autoremove==0.10.0
platformdirs==3.5.1
pluggy==1.3.0
prometheus-client==0.15.0
prompt-toolkit==3.0.32
protobuf==4.25.0
psutil==5.9.4
pure-eval==0.2.2
py4j==0.10.9.7
pycparser==2.21
pycryptodome==3.19.0
pycryptodomex==3.19.0
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
Pygments==2.13.0
pyparsing==3.0.9
pyrsistent==0.19.2
pyspark==3.4.0
pytest==7.4.3
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2022.6
pyunormalize==15.0.0
pywin32==305
pywinpty==2.0.9
pyzmq==24.0.1
regex==2023.10.3
requests==2.28.1
rlp==3.0.0
schemas==0.7.1
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==2.0.23
sqlparse==0.4.3
stack-data==0.6.0
starlette==0.27.0
terminado==0.17.0
tinycss2==1.2.1
tinyec==0.4.0
tomli==2.0.1
toolz==0.12.0
tornado==6.2
traitlets==5.5.0
typing_extensions==4.8.0
tzdata==2022.6
urllib3==1.26.12
uvicorn==0.24.0.post1
virtualenv==20.23.0
wcwidth==0.2.5
web3==6.11.2
webencodings==0.5.1
websocket-client==1.4.2
websockets==12.0
yarl==1.9.2
